Added some input logic in custom game

+++ app/src/main/java/com/example/composition/domain/entity/GameSettings.kt
@@ -7,6 +7,6 @@ import kotlinx.parcelize.Parcelize
 data class GameSettings(
     val maxSumValue: Int,
     val minCountOfRightAnswers: Int,
-    val minPercentOfRightAnswers: Double,
+    val minPercentOfRightAnswers: Int,
     val gameTimeInSeconds: Int,
 ) : Parcelable
